backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
bcrypt>=4.0.1
reportlab>=4.0.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
requests>=2.31.0
email-validator>=2.0.0
openpyxl>=3.1.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
